Move `test_stack_trace` from deprecated test file. (#498)
